+++ packages/cas-validate/src/validate.ts
@@ -13,12 +13,12 @@ export type CasOptions = {
 };
 
 export type CasCallbackExtendedData = {
-	username: string;
-	attributes: Record<string, string[]>;
+	username?: string;
+	attributes?: Record<string, string[]>;
 	// eslint-disable-next-line @typescript-eslint/naming-convention
 	PGTIOU?: string;
-	ticket: string;
-	proxies: string[];
+	ticket?: string;
+	proxies?: string[];
 };
 
 export type CasCallback = (err: any, status?: unknown, username?: string, extended?: CasCallbackExtendedData) => void;