+++ packages/ui-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rocket.chat/ui-client
 
+## 3.0.3
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.3
+
 ## 3.0.2
 
 ### Patch Changes